feat: [cadence-matching] tasklist as shardprocessor (#7435)
<!-- Describe what has changed in this PR -->
**What changed?**
Modelling the shard processor as an extension of tasklist manager.
I am also introducing checks on the parameters needed to generate a new
manager, right now we rely on panic handling in case the some of the
values of the manager are not initializes. This make our unit tests not
able to catch potential panics in production.

<!-- Tell your future self why have you made these changes -->
**Why?**
We want to onboard matching service to shard distributor.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
running the service locally and unit tests

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
low, the new interface will be used in the next pr

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: edigregorio <edigregorio@uber.com>