feat: fix/enforce/address mappers nil handling (#7434)
<!-- Describe what has changed in this PR -->
**What changed?**

This:
- Adds some changes to make mappers nilsafe
- enforces nilaway in lint


<!-- Tell your future self why have you made these changes -->
**Why?**


<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**


<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

This is making a few mappers return nil to values which would previously
of paniced. This is probably not the right solution and they probably
should be changed to return an error. However, thats a larger set of
changes because the risks seem fairly minimal (register domain
specifically)


<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: David Porter <david.porter@uber.com>