fix(BoundedAckCache): fix the ReplicatorCacheMaxSize dynamic config KeyName (#7445)
<!-- Describe what has changed in this PR -->
**What changed?**
Fixed the KeyName to match the ReplicatorCacheMaxSize variable name
Added log to indicate that the budgetManager is being initialized

<!-- Tell your future self why have you made these changes -->
**Why?**
It was confusing to have a KeyName different from the variable used

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Tested locally with the simulation

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: fimanishi <fimanishi@gmail.com>