fix(ReplicationBudgetManager): enable ReplicationBudgetManager by default (#7448)
<!-- Describe what has changed in this PR -->
**What changed?**
Enable ReplicationBudgetManager by default, removing the dynamic config
to enable it.

<!-- Tell your future self why have you made these changes -->
**Why?**
It can be disabled directly in the manager and it's disabled by default.
It also creates confusion and allows the replication cache to be enable
by shard only if the configs are not set properly. This could cause OOM
issues.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
Ran simulation and checked logs locally

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
This enables the budget manager by default but the budget manager
capacity is set to zero. It should be unless the capacity has been
explicitly set to not 0.

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

---------

Signed-off-by: fimanishi <fimanishi@gmail.com>