feat(executor-client): split function for local shard assignment (#7446)
<!-- Describe what has changed in this PR -->
**What changed?**
Instead of having a function that replace the assignment generated with
the local logic we have two different a functions, one to assign a new
shard and the other to remove the assignment. Also these function are
used in the canary to simulate the shard assignment.
The changes to the canary are made to make more similar to the matching
service.


<!-- Tell your future self why have you made these changes -->
**Why?**
we need to onboard matching and to model the local logic with assignment
replacement is adding too much complexity.

<!-- How have you verified this change? Tested locally? Added a unit
test? Checked in staging env? -->
**How did you test it?**
local testing and unit testing

<!-- Assuming the worst case, what can be broken when deploying this
change to production? -->
**Potential risks**
the change is only used in canary service which has been tested

<!-- Is it notable for release? e.g. schema updates, configuration or
data migration required? If so, please mention it, and also update
CHANGELOG.md -->
**Release notes**

<!-- Is there any documentation updates should be made for config,
https://cadenceworkflow.io/docs/operation-guide/setup/ ? If so, please
open an PR in https://github.com/cadence-workflow/cadence-docs -->
**Documentation Changes**

Signed-off-by: edigregorio <edigregorio@uber.com>