feat: using the machines to slowly converge on the right tooling (#7453)
<!-- Describe what has changed in this PR -->
**What changed?**

I am trying to get rid of testify mockery and other deprecated tools and
I'm realising that for generated code, one way to slowly converge on the
right location for is to encode these opinions in the AI instructions.
It's not a real migration, but it will start to make the migration
easier.

I really doubt we'd ever be able to get a complete list of development
opinions agreed upon, but I hope a few uncontroversial ones can make it
in here and we can incrementally improve this as need be.

This is not at all a complete list of development opinions, so please
feel free to suggest / add / increment.

cf https://www.anthropic.com/engineering/claude-code-best-practices

Signed-off-by: David Porter <david.porter@uber.com>